Fix saving (u64 BigInt bug), type-specific edit forms, scrollable chart

+++ src/lib.rs
@@ -42,7 +42,7 @@ impl BabyTracker {
         duration_minutes: Option<u32>,
         notes: Option<String>,
         timestamp: &str,
-    ) -> Result<u64, JsError> {
+    ) -> Result<u32, JsError> {
         self.inner
             .add_feeding(baby_name, feeding_type, amount_ml, duration_minutes, notes, timestamp)
             .map_err(|e| JsError::new(&e))
@@ -51,7 +51,7 @@ impl BabyTracker {
     #[wasm_bindgen(js_name = updateFeeding)]
     pub fn update_feeding(
         &mut self,
-        id: u64,
+        id: u32,
         feeding_type: &str,
         amount_ml: Option<f64>,
         duration_minutes: Option<u32>,
@@ -64,7 +64,7 @@ impl BabyTracker {
     }
 
     #[wasm_bindgen(js_name = deleteFeeding)]
-    pub fn delete_feeding(&mut self, id: u64) -> bool {
+    pub fn delete_feeding(&mut self, id: u32) -> bool {
         self.inner.delete_feeding(id)
     }
 
@@ -77,7 +77,7 @@ impl BabyTracker {
         dejection_type: &str,
         notes: Option<String>,
         timestamp: &str,
-    ) -> Result<u64, JsError> {
+    ) -> Result<u32, JsError> {
         self.inner
             .add_dejection(baby_name, dejection_type, notes, timestamp)
             .map_err(|e| JsError::new(&e))
@@ -86,7 +86,7 @@ impl BabyTracker {
     #[wasm_bindgen(js_name = updateDejection)]
     pub fn update_dejection(
         &mut self,
-        id: u64,
+        id: u32,
         dejection_type: &str,
         notes: Option<String>,
         timestamp: &str,
@@ -97,7 +97,7 @@ impl BabyTracker {
     }
 
     #[wasm_bindgen(js_name = deleteDejection)]
-    pub fn delete_dejection(&mut self, id: u64) -> bool {
+    pub fn delete_dejection(&mut self, id: u32) -> bool {
         self.inner.delete_dejection(id)
     }
 
@@ -110,7 +110,7 @@ impl BabyTracker {
         weight_kg: f64,
         notes: Option<String>,
         timestamp: &str,
-    ) -> Result<u64, JsError> {
+    ) -> Result<u32, JsError> {
         self.inner
             .add_weight(baby_name, weight_kg, notes, timestamp)
             .map_err(|e| JsError::new(&e))
@@ -119,7 +119,7 @@ impl BabyTracker {
     #[wasm_bindgen(js_name = updateWeight)]
     pub fn update_weight(
         &mut self,
-        id: u64,
+        id: u32,
         weight_kg: f64,
         notes: Option<String>,
         timestamp: &str,
@@ -130,7 +130,7 @@ impl BabyTracker {
     }
 
     #[wasm_bindgen(js_name = deleteWeight)]
-    pub fn delete_weight(&mut self, id: u64) -> bool {
+    pub fn delete_weight(&mut self, id: u32) -> bool {
         self.inner.delete_weight(id)
     }
